requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
cffi==1.14.5
cryptography==3.4.6; python_version >= '3.6'
django-rest-knox==4.1.0
django==3.1.7
djangorestframework==3.12.2
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
